Implement recovery helper
Add calculate_progress method
