+++ app/views/recoveries/_recovery_details.html.haml
@@ -1,4 +1,4 @@
-- progress = recovery.target_days.zero? ? 0 : (100 * recovery.current_record) / recovery.target_days
+- progress = calculate_progress(recovery)
 %article.pl-5.py-5.mx-auto.max-w-lg.rounded-md.border.border-2.border-neutral-300.bg-white
   .pr-5
     .flex.items-start.justify-between
